// src/interfaces/matchTable.ts
import { MatchTableResponseEntity } from "./enttiry/matchTable";
import { MatchTablePlayerEntity } from "./enttiry/matchTablePlayer";
import { MatchTableOrderEntity } from "./enttiry/matchTableOrder";
import { MatchTableResultEntity } from "./enttiry/matchTableResult";

export interface MathTableGetRequest {
  match_id: string;
  password: string;
}

export interface MathTableAuthCheckRequest {
  match_id: string;
  auth_password: string;
}

export interface MatchTableCreateRequest {
  title: string;
  match_id: string;
  password: string;
  auth_password?: string;
  player: string[];
}

export interface MatchTableCreate {
  title: string;
  match_id: string;
  password: string;
  auth_password?: string;
}

export type MatchTableData = {
  matchTable: MatchTableResponseEntity;
  matchTablePlayer: MatchTablePlayerEntity[];
  matchTableOrder: MatchTableOrderEntity[];
  matchTableResult: MatchTableResultEntity[];
};

export interface MatchTableUpdateRequest {
  match_id: string;
  auth_password: string | null;
  matchTableOrder: MatchTableOrderEntity[];
  matchTableResult: MatchTableResultEntity[];
}